Fix warnings and missing or extra dependencies.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@4305 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ eklm/geoeklm/include/GeoESTRCreator.h
@@ -88,7 +88,7 @@ namespace Belle2 {
       double* z;
       double* rmin;
       double* rmax;
-      int zsub;
+      double zsub;
       double rminsub;
       double rmaxsub;
       double thick_eiron;
